Fixes for tooltips, t_crops, t_foliage. - Fixed: tooltips sometimes not being updated. - Fixed: t_crops and t_foliage weren't working correctly.
Moved the debugger notification for a failed assertion from the exception catch point to the Assert_Fail function.
Changed m_itEqTradeWindow.m_iPlatinum type. Explanation in the code.
Fix for EF_FixCanSeeInClosedConts and script-created trade windows (used as example by Mordaunt's housing system).
Fixed m_itLight.m_burned data type, which unaligned MOREX/MOREY/MOREZ and broke the access to those properties.

+++ src/network/receive.cpp
@@ -1136,7 +1136,7 @@ bool PacketSecureTradeReq::onReceive(NetState* net)
 		}
 
 		case SECURE_TRADE_UPDATEGOLD:	// update trade window virtual gold
-			container->Trade_UpdateGold(readInt32(), readInt32());
+			container->Trade_UpdateGold((byte)readInt32(), readInt32());
 			return true;
 		default:
 			return true;
